requirements.txt
flask==2.2.2
SQLAlchemy==1.4.41
psycopg2-binary==2.9.3
Flask-Migrate==3.1.0
gunicorn==20.1.0
python-dotenv==0.21.0
redis==4.3.4